:building_construction: exploded the configuration and added more blocks
